Fixing attributes types

+++ src/main/java/com/ia/knn/infrastructure/dto/GridMapping.java
@@ -13,5 +13,5 @@ public class GridMapping {
 
     List<Element> gridElements;
     List<Element> testElements;
-    long kFactor;
+    double kFactor;
 }
